+++ sw.js
@@ -1,4 +1,4 @@
-var CACHE_VERSION = 144;
+var CACHE_VERSION = 147;
 var CACHE_NAME = 'nyhome-v' + CACHE_VERSION;
 
 var APP_SHELL = [